requirements.txt
# Django framework
Django>=5.0

# Django REST Framework
djangorestframework>=3.14.0

# JWT authentication
djangorestframework-simplejwt>=5.3.0

# CORS headers
django-cors-headers>=4.0.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# NumPy - sensor data generation and ML preprocessing
numpy>=1.24.0

# Faker - generating realistic fake data
Faker>=20.0.0

# scikit-learn - ML models (Isolation Forest, etc.)
scikit-learn>=1.3.0

# pandas - data manipulation and time-series analysis
pandas>=2.0.0
